Revert "Fix problem with milliseconds overflowing for long uptime on OSX"
This reverts commit 99afd933e06754a88603996c851ba1729b52cce2.

+++ modules/juce_events/native/juce_mac_MessageManager.mm
@@ -383,7 +383,7 @@ bool MessageManager::runDispatchLoopUntil (int millisecondsToRunFor)
     jassert (millisecondsToRunFor >= 0);
     jassert (isThisTheMessageThread()); // must only be called by the message thread
 
-    auto endTime = (int64) Time::currentTimeMillis() + millisecondsToRunFor;
+    auto endTime = Time::currentTimeMillis() + millisecondsToRunFor;
 
     while (quitMessagePosted.get() == 0)
     {
